+++ src/input_context/context_instance.rs
@@ -1,12 +1,17 @@
 use std::{
     any::{self, TypeId},
     cmp::Ordering,
+    fmt::Debug,
 };
 
-use bevy::{prelude::*, utils::Entry};
+use bevy::{
+    prelude::*,
+    utils::{Entry, HashMap},
+};
 
 use super::{
-    input_action::{Accumulation, ActionData, ActionOutput, ActionsData, InputAction},
+    events::{ActionEvents, Canceled, Completed, Fired, Ongoing, Started},
+    input_action::{Accumulation, ActionOutput, InputAction},
     input_condition::InputCondition,
     input_modifier::{negate::Negate, swizzle_axis::SwizzleAxis, InputModifier},
     trigger_tracker::TriggerTracker,
@@ -14,7 +19,6 @@ use super::{
 use crate::{
     action_value::{ActionValue, ActionValueDim},
     input::{input_reader::InputReader, GamepadDevice, Input},
-    ActionState,
 };
 
 /// Instance for [`InputContext`](super::InputContext).
@@ -358,6 +362,214 @@ impl ActionBind {
             action.trigger_events(commands, entities);
         }
     }
+}
+
+/// Map for actions to their data.
+///
+/// Can be accessed from [`InputCondition::evaluate`]
+/// or [`ContextInstances::get`](super::ContextInstances::get).
+#[derive(Default, Deref, DerefMut)]
+pub struct ActionsData(pub HashMap<TypeId, ActionData>);
+
+impl ActionsData {
+    /// Returns associated state for action `A`.
+    pub fn action<A: InputAction>(&self) -> Option<&ActionData> {
+        self.get(&TypeId::of::<A>())
+    }
+
+    /// Inserts a state for action `A`.
+    ///
+    /// Returns previosly associated state if present.
+    pub fn insert_action<A: InputAction>(&mut self, action: ActionData) -> Option<ActionData> {
+        self.insert(TypeId::of::<A>(), action)
+    }
+}
+
+/// Tracker for action state.
+///
+/// Stored inside [`ActionsData`].
+#[derive(Clone, Copy)]
+pub struct ActionData {
+    state: ActionState,
+    events: ActionEvents,
+    value: ActionValue,
+    elapsed_secs: f32,
+    fired_secs: f32,
+    trigger_events: fn(&Self, &mut Commands, &[Entity]),
+}
+
+impl ActionData {
+    /// Creates a new instance associated with action `A`.
+    ///
+    /// [`Self::trigger_events`] will trigger events for `A`.
+    #[must_use]
+    pub fn new<A: InputAction>() -> Self {
+        Self {
+            state: Default::default(),
+            events: ActionEvents::empty(),
+            value: ActionValue::zero(A::Output::DIM),
+            elapsed_secs: 0.0,
+            fired_secs: 0.0,
+            trigger_events: Self::trigger_events_typed::<A>,
+        }
+    }
+
+    /// Updates internal state.
+    pub fn update(
+        &mut self,
+        time: &Time<Virtual>,
+        state: ActionState,
+        value: impl Into<ActionValue>,
+    ) {
+        match self.state {
+            ActionState::None => {
+                self.elapsed_secs = 0.0;
+                self.fired_secs = 0.0;
+            }
+            ActionState::Ongoing => {
+                self.elapsed_secs += time.delta_seconds();
+                self.fired_secs = 0.0;
+            }
+            ActionState::Fired => {
+                self.elapsed_secs += time.delta_seconds();
+                self.fired_secs += time.delta_seconds();
+            }
+        }
+
+        self.events = ActionEvents::new(self.state, state);
+        self.state = state;
+        self.value = value.into();
+    }
+
+    /// Triggers events resulting from a state transition after [`Self::update`].
+    ///
+    /// See also [`Self::new`].
+    pub fn trigger_events(&self, commands: &mut Commands, entities: &[Entity]) {
+        (self.trigger_events)(self, commands, entities);
+    }
+
+    /// A typed version of [`Self::trigger_events`].
+    fn trigger_events_typed<A: InputAction>(&self, commands: &mut Commands, entities: &[Entity]) {
+        for (_, event) in self.events.iter_names() {
+            match event {
+                ActionEvents::STARTED => {
+                    trigger_for_each(
+                        commands,
+                        entities,
+                        Started::<A> {
+                            value: A::Output::convert_from(self.value),
+                            state: self.state,
+                        },
+                    );
+                }
+                ActionEvents::ONGOING => {
+                    trigger_for_each(
+                        commands,
+                        entities,
+                        Ongoing::<A> {
+                            value: A::Output::convert_from(self.value),
+                            state: self.state,
+                            elapsed_secs: self.elapsed_secs,
+                        },
+                    );
+                }
+                ActionEvents::FIRED => {
+                    trigger_for_each(
+                        commands,
+                        entities,
+                        Fired::<A> {
+                            value: A::Output::convert_from(self.value),
+                            state: self.state,
+                            fired_secs: self.fired_secs,
+                            elapsed_secs: self.elapsed_secs,
+                        },
+                    );
+                }
+                ActionEvents::CANCELED => {
+                    trigger_for_each(
+                        commands,
+                        entities,
+                        Canceled::<A> {
+                            value: A::Output::convert_from(self.value),
+                            state: self.state,
+                            elapsed_secs: self.elapsed_secs,
+                        },
+                    );
+                }
+                ActionEvents::COMPLETED => {
+                    trigger_for_each(
+                        commands,
+                        entities,
+                        Completed::<A> {
+                            value: A::Output::convert_from(self.value),
+                            state: self.state,
+                            fired_secs: self.fired_secs,
+                            elapsed_secs: self.elapsed_secs,
+                        },
+                    );
+                }
+                _ => unreachable!("iteration should yield only named flags"),
+            }
+        }
+    }
+
+    /// Returns the current state.
+    pub fn state(&self) -> ActionState {
+        self.state
+    }
+
+    /// Returns events triggered by a transition of [`Self::state`] since the last update.
+    pub fn events(&self) -> ActionEvents {
+        self.events
+    }
+
+    /// Returns the value since the last update.
+    pub fn value(&self) -> ActionValue {
+        self.value
+    }
+
+    /// Time the action was in [`ActionState::Ongoing`] and [`ActionState::Fired`] states.
+    pub fn elapsed_secs(&self) -> f32 {
+        self.elapsed_secs
+    }
+
+    /// Time the action was in [`ActionState::Fired`] state.
+    pub fn fired_secs(&self) -> f32 {
+        self.fired_secs
+    }
+}
+
+/// Triggers a copyable event for each entity separately and logs it.
+///
+// It's cheaper to copy the event than to clone the entities.
+fn trigger_for_each<E: Event + Debug + Clone + Copy>(
+    commands: &mut Commands,
+    entities: &[Entity],
+    event: E,
+) {
+    for &entity in entities {
+        trace!("triggering `{event:?}` for `{entity}`");
+        commands.trigger_targets(event, entity);
+    }
+}
+
+/// State for [`ActionData`].
+///
+/// States are ordered by their significance.
+///
+/// See also [`ActionEvents`].
+#[derive(Clone, Copy, Default, Debug, PartialEq, Eq, PartialOrd, Ord)]
+pub enum ActionState {
+    /// Condition is not triggered.
+    #[default]
+    None,
+    /// Condition has started triggering, but has not yet finished.
+    ///
+    /// For example, [`Hold`](super::input_condition::hold::Hold) condition
+    /// requires its state to be maintained over several frames.
+    Ongoing,
+    /// The condition has been met.
+    Fired,
 }
 
 /// Associated input for [`ActionBind`].