+++ src/lib.rs
@@ -72,10 +72,11 @@ pub mod prelude {
         input::{GamepadDevice, Input, InputModKeys, ModKeys},
         input_context::{
             context_instance::{
-                ActionBind, ContextInstance, GamepadStick, InputBind, InputBindModCond,
+                ActionBind, ActionData, ActionState, ContextInstance, GamepadStick, InputBind,
+                InputBindModCond,
             },
             events::*,
-            input_action::{Accumulation, ActionData, ActionState, InputAction},
+            input_action::{Accumulation, InputAction},
             input_condition::{
                 block_by::*, chord::*, condition_timer::*, hold::*, hold_and_release::*,
                 just_press::*, press::*, pulse::*, release::*, tap::*, ConditionKind,